src/backend/requirements.txt
fastapi[standard]
requests
uvicorn 
opencv-python
pillow
numpy
torch --index-url https://download.pytorch.org/whl/cpu
--pre torch torchvision torchaudio --extra-index-url https://download.pytorch.org/whl/nightly/cpu
torch
torchvision
torchaudio
ultralytics
speechrecognition 
gtts